requirements.txt
openai==1.35.0
anthropic==0.28.0
python-dotenv==1.0.0
pydantic==2.7.0
rich==13.7.0
click==8.1.7
fastapi==0.111.0
pytest==8.2.0